requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
h11==0.14.0
idna==3.4
oauthlib==3.2.2
outcome==1.2.0
Pillow==9.3.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.12.1
urllib3==1.26.12
wsproto==1.2.0
